Features/ManageabilityPkg: edk2 migration CI fix - CodeQL
This is the prerequisite of migrate Manageability package
from edk2-platforms to edk2 repository.

Fix the CodeQL errors.

+++ Features/ManageabilityPkg/Library/ManageabilityTransportSerialLib/Common/SerialCommon.c
@@ -219,7 +219,7 @@ SerialTransportWrite (
   UINT8               *Request;
   UINT32              RequestLength;
   UINT8               RetryCount;
-  UINT8               Index;
+  UINT32              Index;
   UINT8               Character;
   UINT8               EscapedCharacterCount;
   UINT8               NetFunction;
